requirements.txt
streamlit
requests
openai
whisper
python-dotenv
langchain
langchain_community
sounddevice